De-abstractify InstanceListCommand (#360)
